on affiche un degradey.

+++ srcs/parser/f_grad_parse.c
@@ -6,7 +6,7 @@
 /*   By: kdouveno <kdouveno@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2018/10/08 16:15:27 by kdouveno          #+#    #+#             */
-/*   Updated: 2018/10/18 17:38:46 by gperez           ###   ########.fr       */
+/*   Updated: 2018/11/05 17:04:56 by kdouveno         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -22,12 +22,12 @@ static int		check_arg(t_grad *grad, char *l1, char *l2)
 	}
 	if (!(ft_strcmp(l1, "color1")))
 	{
-		grad->color1 = (t_color)ft_atoi(l2);
+		grad->color1 = (t_color)(unsigned int)ft_atoi(l2);
 		return (0);
 	}
 	if (!(ft_strcmp(l1, "color2")))
 	{
-		grad->color2 = (t_color)ft_atoi(l2);
+		grad->color2 = (t_color)(unsigned int)ft_atoi(l2);
 		return (0);
 	}
 	return (1);
@@ -62,7 +62,7 @@ t_grad			*grad_parse_2(t_env *e, int fd)
 	if (!(grad = malloc(sizeof(t_grad))))
 		error(e, MALLOC_ERROR);
 	*grad = (t_grad){(t_pt){0, 0, 0}, (t_vec){0, 0, 0}, 0,
-		(t_color)0, (t_color)0, NULL};
+		(t_color)(unsigned int)0, (t_color)(unsigned int)0, NULL};
 	while ((res = get_next_line(fd, &line)) > 0
 		&& get_prop(e, line, &l1, &l2) != 1)
 	{
